+++ ChangeLog.md
@@ -5,6 +5,7 @@
 
 ## 1.0
 
+- FIX : TOKEN compat v16 - *24/06/2022* - 1.1.3  
 - FIX: Compatibility V13+V14 - GETPOST; module descriptor updated - *30/06/2021* - 1.1.1
 - FIX: Compatibility V13 - delete box folder - *18/05/2021* - 1.1.1
 - Initial release